requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
blobfile==3.0.0
boto3==1.35.29
boto3-stubs==1.34.160
botocore==1.35.29
botocore-stubs==1.34.160
Brotli==1.1.0
brotlicffi==1.1.0.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.4
fastapi-cli==0.0.5
filelock==3.15.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.0
greenlet==3.0.3
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface_hub==0.25.2
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-community==0.3.1
langchain-core==0.3.7
langchain-openai==0.2.1
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.129
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-boto3-dynamodb==1.35.24
mypy-boto3-s3==1.35.42
mypy_extensions==1.0.0
nest_asyncio==1.6.0
numpy==1.26.4
olefile==0.47
openai==1.52.2
orjson==3.10.6
packaging==24.1
pandas==2.2.2
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==2.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
podman-compose==1.2.0
prompt_toolkit==3.0.47
propcache==0.2.0
psutil==6.0.0
psycopg==3.2.2
psycopg-binary==3.2.2
ptyprocess==0.7.0
pure_eval==0.2.2
pyasn1==0.6.0
pycares==4.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pypdf==5.0.0
pyspnego==0.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
redis==5.2.0
regex==2024.9.11
requests==2.32.3
rich==13.7.1
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
smbprotocol==1.12.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sse-starlette==2.1.3
stack_data==0.6.3
starlette==0.40.0
tenacity==8.4.1
threadpoolctl==3.4.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
types-awscrt==0.21.5
types-pyOpenSSL==24.1.0.20240425
types-redis==4.6.0.20240425
types-s3transfer==0.10.1
typing_extensions==4.12.2
typing_inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
yarl==1.17.0
